refactor: exported GalaxyMap app from index file
